# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MeadTools API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for MeadTools public, authenticated user, and device integration endpoints."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current MeadTools host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Send the access token in the Authorization header as `Bearer <token>`. External API clients should use MeadTools access tokens returned by the auth endpoints."
      }
    },
    "schemas": {
...
      "RecipeDerivedStateResponseBody": {
        "type": "object",
        "properties": {
          "recipeData": {
            "$ref": "#/components/schemas/recipeDataV2Schema"
          },
          "derived": {
            "$ref": "#/components/schemas/RecipeDerivedStateResponse"
          }
        },
        "required": [
          "recipeData",
          "derived"
        ]
      },
...
      "RecipeDerivedFailureErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Failed to calculate recipe derived state",
              "An error occurred while fetching the recipe"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RecipeDataV2Response": {
        "$ref": "#/components/schemas/recipeDataV2Schema"
      },
      "RecipeDerivedValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Invalid recipe data payload."
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "nutrientAmountsByKeyNumberSchema": {
        "type": "object",
        "properties": {
          "fermO": {
            "type": "number"
          },
          "fermK": {
            "type": "number"
          },
          "dap": {
            "type": "number"
          },
          "other": {
            "type": "number"
          }
        },
        "required": [
          "fermO",
          "fermK",
          "dap",
          "other"
        ]
      },
...
      "RecipeDerivedStateResponse": {
        "type": "object",
        "properties": {
          "gravity": {
            "type": "object",
            "properties": {
              "ogPrimary": {
                "type": "number"
              },
              "backsweetenedFg": {
                "type": "number"
              },
              "totalForAbv": {
                "type": "number"
              }
            },
            "required": [
              "ogPrimary",
              "backsweetenedFg",
              "totalForAbv"
            ]
          },
          "volume": {
            "type": "object",
            "properties": {
              "unit": {
                "$ref": "#/components/schemas/VolumeUnit"
              },
              "primary": {
                "type": "number"
              },
              "secondary": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "primaryL": {
                "type": "number"
              },
              "secondaryL": {
                "type": "number"
              },
              "totalL": {
                "type": "number"
              }
            },
            "required": [
              "unit",
              "primary",
              "secondary",
              "total",
              "primaryL",
              "secondaryL",
              "totalL"
            ]
          },
          "alcohol": {
            "type": "object",
            "properties": {
              "abv": {
                "type": "number"
              },
              "delle": {
                "type": "number"
              }
            },
            "required": [
              "abv",
              "delle"
            ]
          },
          "stabilizers": {
            "type": "object",
            "properties": {
              "sorbate": {
                "type": "number"
              },
              "sulfite": {
                "type": "number"
              },
              "campden": {
                "type": "number"
              }
            },
            "required": [
              "sorbate",
              "sulfite",
              "campden"
            ]
          },
          "nutrients": {
            "type": "object",
            "properties": {
              "targetYanPpm": {
                "type": "number"
              },
              "remainingYanPpm": {
                "type": "number"
              },
              "numberOfAdditions": {
                "type": "number"
              },
              "nutrientAdditions": {
                "type": "object",
                "properties": {
                  "totalGrams": {
                    "$ref": "#/components/schemas/nutrientAmountsByKeyNumberSchema"
                  },
                  "perAddition": {
                    "$ref": "#/components/schemas/nutrientAmountsByKeyNumberSchema"
                  }
                },
                "required": [
                  "totalGrams",
                  "perAddition"
                ]
              },
              "providedYanPpm": {
                "$ref": "#/components/schemas/nutrientAmountsByKeyNumberSchema"
              },
              "goFerm": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "water": {
                    "type": "number"
                  }
                },
                "required": [
                  "amount",
                  "water"
                ]
              }
            },
            "required": [
              "targetYanPpm",
              "remainingYanPpm",
              "numberOfAdditions",
              "nutrientAdditions",
              "providedYanPpm",
              "goFerm"
            ]
          }
        },
        "required": [
          "gravity",
          "volume",
          "alcohol",
          "stabilizers",
          "nutrients"
        ]
      },
...
      "VolumeUnit": {
        "$ref": "#/components/schemas/sharedVolumeUnitSchema"
      },
...
    },
...
  },
  "paths": {
...
    "/recipes/derived": {
      "post": {
        "operationId": "post-recipes-derived",
        "summary": "Calculate recipe derived state",
        "description": "Calculates the same derived recipe state used by the recipe builder from a v2 recipe data payload.",
        "tags": [
          "Recipes"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeDataV2Response"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDerivedStateResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDerivedValidationErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDerivedFailureErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}